Move `contract` to `contract-factory`
