+++ packages/program/src/contract-factory/index.ts
@@ -1,0 +1,11 @@
+export {
+  default as ContractFactory,
+  DeployContractOptions,
+  DeployContractResult,
+} from './contract-factory';
+export {
+  getContractRoot,
+  getContractStorageRoot,
+  getContractId,
+  hexlifyWithPrefix,
+} from './utils';
