+++ packages/program/src/test-utils/launch-test-node.ts
@@ -10,7 +10,10 @@ import { readFileSync } from 'fs';
 import * as path from 'path';
 import { mergeDeepRight } from 'ramda';
 
-import type { DeployContractOptions, DeployContractResult } from '../contract/contract-factory';
+import type {
+  DeployContractOptions,
+  DeployContractResult,
+} from '../contract-factory';
 
 export interface DeployableContractFactory {
   deploy(wallet: Account, options?: DeployContractOptions): Promise<DeployContractResult>;
